Refactor: MemberRepositoryTest 에서 MemberScrapRepositoryTest 분리
