+++ src/test/java/net/mureng/mureng/annotation/MurengDataTest.java
@@ -9,6 +9,7 @@ import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.TestExecutionListeners;
 import org.springframework.test.context.support.DependencyInjectionTestExecutionListener;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -24,5 +25,6 @@ import java.lang.annotation.Target;
 @TestExecutionListeners({ DependencyInjectionTestExecutionListener.class,
         TransactionDbUnitTestExecutionListener.class })
 @DbUnitConfiguration(databaseConnection = "dbUnitDatabaseConnection")
+@Transactional
 public @interface MurengDataTest {
 }
